requirements.txt
argparse==1.4.0
DateTime==4.3
python-dotenv==0.19.2
requests==2.26.0
psycopg2==2.9.2
pymssql==2.2.2
SQLAlchemy==1.4.27
configparser==5.2.0
PyMySQL==1.0.2
pandas==1.3.4
mongoengine==0.23.1
pymongo==3.11.3
pymongo[srv]
numpy==1.21.4
faker==9.9.0
faker_airtravel==0.4
faker-vehicle==0.2.0
Flask==2.0.2
minio==7.1.2
cassandra-driver==3.25.0
confluent-kafka==1.7.0
confluent_avro==1.8.0
avro-python3==1.10.2 